Add batch size and delay docs
This commit adds documentation for batch size and delay to the
DB.Batch() call. Previously there was no reference to these properties
from the Batch() call so it was nearly impossible for anyone to
know to adjust these settings.

Thanks to Kelly Sommers for bringing the doc issue to my attention.